mv requires_dev_feature_enabled to utils
